refactor: Rename/move components, remove dead code.
