+++ lib/components/app/responsive-webapp.js
@@ -21,7 +21,7 @@ import { AUTH0_AUDIENCE, AUTH0_SCOPE, URL_ROOT } from '../../util/constants'
 import { getActiveItinerary, getTitle } from '../../util/state'
 import AfterSignInScreen from '../user/after-signin-screen'
 import BeforeSignInScreen from '../user/before-signin-screen'
-import FavoritePlaceScreen from '../user/favorite-locations/favorite-place-screen'
+import FavoritePlacePage from '../user/places/favorite-place-page'
 import SavedTripList from '../user/saved-trip-list'
 import SavedTripScreen from '../user/saved-trip-screen'
 import UserAccountScreen from '../user/user-account-screen'
@@ -212,6 +212,7 @@ class RouterWrapperWithAuth0 extends Component {
         history={history}>
         <div>
           <Switch>
+            {/* TODO: Review the url schemes for consistency and brevity. */}
             <Route
               exact
               path={[
@@ -245,7 +246,7 @@ class RouterWrapperWithAuth0 extends Component {
               path={'/places/:id'}
               component={(routerProps) => {
                 const props = this._combineProps(routerProps)
-                return <FavoritePlaceScreen {...props} />
+                return <FavoritePlacePage {...props} />
               }}
             />
             <Route
