+++ lib/components/user/places/favorite-place-page.js
@@ -1,15 +1,8 @@
 import { withAuthenticationRequired } from '@auth0/auth0-react'
 import clone from 'clone'
-import { Field, Form, Formik } from 'formik'
+import { Form, Formik } from 'formik'
 import React, { Component } from 'react'
-import {
-  Button,
-  FormControl,
-  FormGroup,
-  HelpBlock,
-  ToggleButton,
-  ToggleButtonGroup
-} from 'react-bootstrap'
+import { Button } from 'react-bootstrap'
 import { connect } from 'react-redux'
 import styled from 'styled-components'
 import * as yup from 'yup'
@@ -17,14 +10,11 @@ import * as yup from 'yup'
 import * as uiActions from '../../../actions/ui'
 import * as userActions from '../../../actions/user'
 import DesktopNav from '../../app/desktop-nav'
-import {
-  FavoriteLocationField,
-  makeLocationFieldLocation
-} from './favorite-location-controls'
 import Icon from '../../narrative/icon'
-import { getErrorStates, RETURN_TO_CURRENT_ROUTE } from '../../../util/ui'
+import { RETURN_TO_CURRENT_ROUTE } from '../../../util/ui'
 import { cloneForEditing, isHomeOrWork, PLACE_TYPES } from '../../../util/user'
 import withLoggedInUserSupport from '../with-logged-in-user-support'
+import PlaceEditor from './place-editor'
 
 // Styled components
 const SaveButton = styled(Button)`
@@ -33,17 +23,6 @@ const SaveButton = styled(Button)`
 const BackButton = styled(Button)`
   padding: 0;
 `
-const LargeIcon = styled(Icon)`
-  font-size: 150%;
-`
-const FixedPlaceIcon = styled(LargeIcon)`
-  margin-right: 10px;
-  padding-top: 6px;
-`
-
-const MESSAGES = {
-  SET_PLACE_NAME: 'Set place name'
-}
 
 // Custom places without home or work
 const customPlaces = clone(PLACE_TYPES)
@@ -63,101 +42,10 @@ const validationSchemaShape = {
 }
 
 /**
- * Contains the fields for editing a favorite place.
- * This component uses Formik props that are passed
- * within the Formik context set up by FavoritePlaceScreen.
+ * Lets the user edit the details (address, type, nickname)
+ * of a new or existing favorite place.
  */
-class FavoritePlaceEditor extends Component {
-  _handleLocationChange = ({ location }) => {
-    const { setValues, values } = this.props
-    const { lat, lon, name } = location
-    setValues({
-      ...values,
-      address: name,
-      lat,
-      lon
-    })
-  }
-
-  render () {
-    const { errors, handleBlur, handleChange, values: place } = this.props
-    const isFixed = isHomeOrWork(place)
-    const errorStates = getErrorStates(this.props)
-
-    return (
-      <div style={{marginTop: '40px'}}>
-        {!isFixed && (
-        <>
-          <FormGroup validationState={errorStates.name} >
-            {/* onBlur, onChange, and value are passed automatically. */}
-            <Field as={FormControl}
-              aria-label={MESSAGES.SET_PLACE_NAME}
-              name='name'
-              placeholder={MESSAGES.SET_PLACE_NAME}
-            />
-            <FormControl.Feedback />
-            {/* TODO: Validate/m ake sure names are not duplicate */}
-            {errors.name && <HelpBlock>{errors.name}</HelpBlock>}
-          </FormGroup>
-
-          <FormGroup>
-            <ToggleButtonGroup
-              name='type'
-              type='radio'
-              defaultValue={place.type}
-            >
-              {Object.values(customPlaces).map(({ icon, text, type }) => (
-                <ToggleButton
-                  key={type}
-                  // onBlur and onChange have to be set on individual controls instead of the control group
-                  // in order for Formik to correcly process the changes.
-                  onBlur={handleBlur}
-                  onChange={handleChange}
-                  title={text}
-                  value={type}
-                >
-                  <LargeIcon name={icon} />
-                </ToggleButton>
-              ))}
-            </ToggleButtonGroup>
-          </FormGroup>
-        </>
-        )}
-
-        <div style={{ display: 'flex' }}>
-          {/* For fixed locations, just show the icon for place type instead of all inputs and selectors */}
-          {isFixed && <FixedPlaceIcon name={place.icon} />}
-
-          <FormGroup style={{flexGrow: 1}} validationState={errorStates.address}>
-            <FavoriteLocationField // TODO: mobile version.
-              className='form-control'
-              inputPlaceholder={isFixed ? `Search for ${place.name} location` : 'Search for location'}
-              location={makeLocationFieldLocation(place)}
-              locationType='to'
-              onLocationSelected={this._handleLocationChange}
-              showClearButton={false}
-            />
-            <FormControl.Feedback />
-            {errors.address && <HelpBlock>{errors.address}</HelpBlock>}
-
-            {/* TODO: Implement setting location using map. */}
-          </FormGroup>
-        </div>
-
-        <FormGroup>
-          {/* TODO: Remove */}
-          <p>DEV NOTE: Changes take effect immediately after clicking Save.</p>
-        </FormGroup>
-      </div>
-    )
-  }
-}
-
-/**
- * Renders a user favorite location,
- * and lets the user edit the details (address, type, nickname) of the location.
- */
-class FavoritePlaceScreen extends Component {
+class FavoritePlacePage extends Component {
   /**
    * Navigates back to the previous browser location,
    * whether it is the accounts page or another page that linked to the current URL.
@@ -171,14 +59,14 @@ class FavoritePlaceScreen extends Component {
    * Silently save the changes to the loggedInUser, and go to the previous URL.
    */
   _handleSave = async placeToSave => {
-    // Update the place icon for the location type.
+    // Update the icon for the place type.
     placeToSave.icon = PLACE_TYPES[placeToSave.type].icon
 
     // Save changes to loggedInUser
     const { createOrUpdateUser, loggedInUser, match } = this.props
 
     // Get the places as shown (and not as retrieved from db), so that the index passed from URL applies
-    // (indexes 0 and 1 are for Home and Work locations).
+    // (indexes 0 and 1 are for 'Home' and 'Work' places).
     const clonedUser = cloneForEditing(loggedInUser)
     const placeIndex = match.params.id
 
@@ -287,7 +175,7 @@ class FavoritePlaceScreen extends Component {
                   </div>
 
                   {place
-                    ? <FavoritePlaceEditor {...props} />
+                    ? <PlaceEditor {...props} />
                     : <p>Sorry, the requested place was not found.</p>
                   }
                 </Form>
@@ -315,7 +203,7 @@ const mapDispatchToProps = {
 
 export default withLoggedInUserSupport(
   withAuthenticationRequired(
-    connect(mapStateToProps, mapDispatchToProps)(FavoritePlaceScreen),
+    connect(mapStateToProps, mapDispatchToProps)(FavoritePlacePage),
     RETURN_TO_CURRENT_ROUTE
   ),
   true
