+++ lib/components/user/places/favorite-place-row.js
@@ -1,4 +1,5 @@
 import clone from 'clone'
+import PropTypes from 'prop-types'
 import React, { Component } from 'react'
 import { Button } from 'react-bootstrap'
 import { connect } from 'react-redux'
@@ -7,8 +8,7 @@ import styled, { css } from 'styled-components'
 import * as uiActions from '../../../actions/ui'
 import Icon from '../../narrative/icon'
 
-export const FIELD_HEIGHT_PX = '60px'
-export const ICON_WIDTH_PX = '30px'
+const FIELD_HEIGHT_PX = '60px'
 
 const Container = styled.div`
   align-items: stretch;
@@ -17,7 +17,7 @@ const Container = styled.div`
   margin-bottom: 10px;
 `
 
-const LocationButton = styled(Button)`
+const PlaceButton = styled(Button)`
   flex: 1 0 0;
   overflow: hidden;
   text-align: left;
@@ -27,12 +27,12 @@ const LocationButton = styled(Button)`
   }
 `
 
-const StyledIcon = styled(Icon)`
+const GreyIcon = styled(Icon)`
   color: #888;
   margin-right: 10px;
 `
 
-const LocationContent = styled.span`
+const PlaceContent = styled.span`
   display: inline-block;
 
   & * {
@@ -68,71 +68,89 @@ const MESSAGES = {
  * and lets the user delete the place.
  */
 class FavoritePlaceRow extends Component {
-  // TODO: add prop types.
+  static propTypes = {
+    // A Formik arrayHelper object.
+    arrayHelpers: PropTypes.shape({
+      remove: PropTypes.func.isRequired,
+      replace: PropTypes.func.isRequired
+    }),
+    // The index of the place to edit or delete.
+    index: PropTypes.number,
+    // Whether the place is fixed (e.g. 'Home', 'Work' are fixed.)
+    isFixed: PropTypes.bool,
+    // The place to render.
+    place: PropTypes.shape({
+      address: PropTypes.string,
+      icon: PropTypes.string.isRequired,
+      name: PropTypes.string.isRequired,
+      type: PropTypes.string.isRequired
+    })
+  }
 
   /**
    * Route to /places/<index> for editing an existing place,
    * or to /places/new for creating a new one.
    */
   _handleClick = () => {
-    const { index, location, routeTo } = this.props
-    const placeId = location ? index : 'new'
+    const { index, place, routeTo } = this.props
+    const placeId = place ? index : 'new'
     routeTo(`/places/${placeId}`)
   }
 
   _handleDelete = () => {
-    const { arrayHelpers, index, isFixed, location } = this.props
+    const { arrayHelpers, index, isFixed, place } = this.props
     if (isFixed) {
-      // Just clear the address if location is fixed.
-      const newLocation = clone(location)
-      newLocation.address = ''
-      arrayHelpers.replace(index, newLocation)
+      // Just clear the address if place is fixed.
+      const newPlace = clone(place)
+      newPlace.address = ''
+      arrayHelpers.replace(index, newPlace)
     } else {
       arrayHelpers.remove(index)
     }
   }
 
   render () {
-    const { isFixed, location } = this.props
+    const { isFixed, place } = this.props
 
     let contents
-    if (location) {
+    if (place) {
+      const { address, icon, name, type } = place
       contents = (
         <>
-          <LocationButton
+          <PlaceButton
             aria-label={MESSAGES.EDIT}
             onClick={this._handleClick}
             title={MESSAGES.EDIT}
           >
-            <StyledIcon name={location.icon} size='2x' />
-            <LocationContent>
-              {location.name && <span>{location.name}</span>}
-              <span>{location.address || `Set your ${location.type} address`}</span>
-            </LocationContent>
-          </LocationButton>
+            <GreyIcon name={icon} size='2x' />
+            <PlaceContent>
+              {name && <span>{name}</span>}
+              <span>{address || `Set your ${type} address`}</span>
+            </PlaceContent>
+          </PlaceButton>
 
-          {/* For fixed locations, show Delete only if an address has been provided. */}
-          {(!isFixed || location.address)
+          {/* For fixed places, show Delete only if an address has been provided. */}
+          {(!isFixed || address)
             ? (
               <DeleteButton
                 aria-label={MESSAGES.DELETE}
                 onClick={this._handleDelete}
                 title={MESSAGES.DELETE}
               >
-                <StyledIcon name='trash-o' size='2x' />
+                <GreyIcon name='trash-o' size='2x' />
               </DeleteButton>
             )
             : <DeletePlaceholder />}
         </>
       )
     } else {
-      // If no location is passed, render the Add place button instead.
+      // If no place is passed, render the Add place button instead.
       contents = (
         <>
-          <LocationButton onClick={this._handleClick}>
-            <StyledIcon name='plus' size='2x' />
+          <PlaceButton onClick={this._handleClick}>
+            <GreyIcon name='plus' size='2x' />
             Add another place
-          </LocationButton>
+          </PlaceButton>
           <DeletePlaceholder />
         </>
       )
