+++ lib/components/user/places/favorite-places-pane.js
@@ -25,7 +25,7 @@ const FavoriteLocationsPane = ({ values: userData }) => {
                 index={index}
                 isFixed={isHomeOrWork(loc)}
                 key={index}
-                location={loc}
+                place={loc}
               />
             ))}
