+++ lib/components/user/user-account-screen.js
@@ -11,9 +11,9 @@ import { cloneForEditing, isNewUser } from '../../util/user'
 import DesktopNav from '../app/desktop-nav'
 import AccountSetupFinishPane from './account-setup-finish-pane'
 import ExistingAccountDisplay from './existing-account-display'
-import FavoriteLocationsPane from './favorite-locations/favorite-locations-pane'
 import NewAccountWizard from './new-account-wizard'
 import NotificationPrefsPane from './notification-prefs-pane'
+import FavoritePlacesPane from './places/favorite-places-pane'
 import TermsOfUsePane from './terms-of-use-pane'
 import VerifyEmailScreen from './verify-email-screen'
 import withLoggedInUserSupport from './with-logged-in-user-support'
@@ -84,7 +84,7 @@ class UserAccountScreen extends Component {
   _panes = {
     terms: TermsOfUsePane,
     notifications: NotificationPrefsPane,
-    locations: FavoriteLocationsPane,
+    locations: FavoritePlacesPane,
     finish: AccountSetupFinishPane
   }
